refactor: move maybe_empty to models module
